split history tabs
